pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scene-wiki"
version = "0.1.0"
description = "Generate an Obsidian-style Quartz wiki from a Substack archive."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "openai>=1.68.2",
  "pydantic>=2.10.6",
  "python-dotenv>=1.0.1",
  "typer>=0.15.1",
]

[project.scripts]
scene-wiki = "scene_wiki.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
